refactor: 카테고리 대표 장소 controller와 service를 카테고리 controller와 service로 병합
