+++ src/main/java/courseitda/workspace/application/CategoryService.java
@@ -4,16 +4,21 @@ import courseitda.common.exception.BusinessException;
 import courseitda.common.exception.ErrorCode;
 import courseitda.workspace.application.dto.request.CreateCategoryCommand;
 import courseitda.workspace.application.dto.request.DeleteCategoryCommand;
+import courseitda.workspace.application.dto.request.DeleteRepresentativeCategoryPlaceCommand;
 import courseitda.workspace.application.dto.request.FindAllCategoriesCommand;
 import courseitda.workspace.application.dto.request.FindCategoryCommand;
 import courseitda.workspace.application.dto.request.UpdateCategoryCommand;
 import courseitda.workspace.application.dto.request.UpdateCategorySequenceCommand;
+import courseitda.workspace.application.dto.request.UpdateRepresentativeCategoryPlaceCommand;
 import courseitda.workspace.application.dto.response.CreateCategoryResult;
 import courseitda.workspace.application.dto.response.FindAllCategoriesResult;
 import courseitda.workspace.application.dto.response.FindCategoryResult;
 import courseitda.workspace.application.dto.response.UpdateCategoryResult;
 import courseitda.workspace.application.dto.response.UpdateCategorySequenceResult;
+import courseitda.workspace.application.dto.response.UpdateRepresentativeCategoryPlaceResult;
 import courseitda.workspace.domain.Category;
+import courseitda.workspace.domain.CategoryPlace;
+import courseitda.workspace.domain.CategoryPlaceRepository;
 import courseitda.workspace.domain.CategoryRepository;
 import courseitda.workspace.domain.Workspace;
 import courseitda.workspace.domain.WorkspaceRepository;
@@ -28,8 +33,9 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor
 public class CategoryService {
 
-    private final CategoryRepository categoryRepository;
     private final WorkspaceRepository workspaceRepository;
+    private final CategoryRepository categoryRepository;
+    private final CategoryPlaceRepository categoryPlaceRepository;
 
     @Transactional
     public CreateCategoryResult createCategory(
@@ -44,6 +50,31 @@ public class CategoryService {
         final var savedCategory = categoryRepository.save(category);
 
         return CreateCategoryResult.from(savedCategory);
+    }
+
+    @Transactional
+    public UpdateCategoryResult updateCategory(
+            final UpdateCategoryCommand command
+    ) {
+        final var category = getCategoryById(command.categoryId());
+        category.validateOwnership(command.memberAuthInfo().id());
+
+        category.updateNameAndColor(command.name(), command.color());
+        return UpdateCategoryResult.from(category);
+    }
+
+    @Transactional
+    public UpdateRepresentativeCategoryPlaceResult updateRepresentativeCategoryPlace(
+            final UpdateRepresentativeCategoryPlaceCommand command
+    ) {
+        final var category = getCategoryById(command.categoryId());
+        category.validateOwnership(command.memberAuthInfo().id());
+
+        final var candidatePlace = getCategoryPlaceById(command.categoryPlaceId());
+
+        category.updateRepresentativePlaceTo(candidatePlace);
+
+        return UpdateRepresentativeCategoryPlaceResult.from(category.getRepresentativePlace());
     }
 
     @Transactional
@@ -82,14 +113,11 @@ public class CategoryService {
     }
 
     @Transactional
-    public UpdateCategoryResult updateCategory(
-            final UpdateCategoryCommand command
-    ) {
+    public void deleteRepresentativeCategoryPlace(final DeleteRepresentativeCategoryPlaceCommand command) {
         final var category = getCategoryById(command.categoryId());
         category.validateOwnership(command.memberAuthInfo().id());
 
-        category.updateNameAndColor(command.name(), command.color());
-        return UpdateCategoryResult.from(category);
+        category.updateRepresentativePlaceTo(null);
     }
 
     @Transactional
@@ -123,11 +151,6 @@ public class CategoryService {
         return FindAllCategoriesResult.from(categories);
     }
 
-    private Workspace getWorkspaceByIdentifier(final String workspaceIdentifier) {
-        return workspaceRepository.findByIdentifier(workspaceIdentifier)
-                .orElseThrow(() -> new BusinessException(ErrorCode.WORKSPACE_NOT_FOUND));
-    }
-
     private void validateAllCategoriesExist(final List<Category> categories, final List<Long> categoryIds) {
         if (categories.size() != categoryIds.size()) {
             throw new BusinessException(ErrorCode.PARTIAL_CATEGORY_NOT_FOUND);
@@ -155,8 +178,18 @@ public class CategoryService {
         }
     }
 
+    private Workspace getWorkspaceByIdentifier(final String workspaceIdentifier) {
+        return workspaceRepository.findByIdentifier(workspaceIdentifier)
+                .orElseThrow(() -> new BusinessException(ErrorCode.WORKSPACE_NOT_FOUND));
+    }
+
     private Category getCategoryById(final Long categoryId) {
         return categoryRepository.findById(categoryId)
                 .orElseThrow(() -> new BusinessException(ErrorCode.CATEGORY_NOT_FOUND));
+    }
+
+    private CategoryPlace getCategoryPlaceById(final Long categoryPlaceId) {
+        return categoryPlaceRepository.findById(categoryPlaceId)
+                .orElseThrow(() -> new BusinessException(ErrorCode.CATEGORY_PLACE_NOT_FOUND));
     }
 }
